backend/requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
altgraph==0.17.4
arcgis==2.4.0
attrs==25.1.0
backports.tarfile==1.2.0
blinker==1.9.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
cloudpickle==3.1.0
colorama==0.4.6
contourpy==1.3.1
cryptography==44.0.0
cycler==0.12.1
dask==2024.12.0
decorator==5.1.1
Flask==3.1.0
Flask-Cors==5.0.0
fonttools==4.55.2
frozenlist==1.5.0
fsspec==2024.10.0
geographiclib==2.0
geojson==3.1.0
geomet==1.1.0
geopy==2.4.1
googlemaps==4.10.0
gssapi==1.9.0
idna==3.10
importlib_metadata==8.5.0
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
Jinja2==3.1.4
keyring==25.5.0
kiwisolver==1.4.7
load-dotenv==0.1.0
locket==1.0.0
lxml==5.3.0
MarkupSafe==3.0.2
matplotlib==3.9.3
matplotlib-inline==0.1.7
more-itertools==10.5.0
multidict==6.1.0
numpy==1.26.4
oauthlib==3.2.2
packaging==24.2
paho-mqtt==2.1.0
pandas==2.1.4
partd==1.4.2
pefile==2023.2.7
pillow==11.0.0
propcache==0.2.1
puremagic==1.28
pyarrow==18.1.0
pycparser==2.22
pyinstaller==6.12.0
pyinstaller-hooks-contrib==2025.1
pylerc==4.0
pyodbc==5.2.0
pyparsing==3.2.0
pyproj==3.7.0
pyshp==2.3.1
pyspnego==0.11.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
pywin32==308
pywin32-ctypes==0.2.3
PyYAML==6.0.2
requests==2.31.0
requests-gssapi==1.3.0
requests-kerberos==0.15.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
shapely==2.0.6
six==1.16.0
sspilib==0.2.0
toolz==1.0.0
traitlets==5.14.3
truststore==0.10.0
tzdata==2024.2
ujson==5.10.0
urllib3==2.2.3
waitress==3.0.2
Werkzeug==3.1.3
yarl==1.18.3
zipp==3.21.0
